requirement.txt
amqp==2.6.1
appdirs==1.4.4
Babel==2.8.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bootstrap-admin==0.4.4
celery==4.4.7
certifi==2019.11.28
cffi==1.14.0
cfgv==2.0.1
chardet==3.0.4
click==7.1.2
contextvars==2.4
cryptography==2.9.2
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.1
dj-database-url==0.5.0
Django==2.0
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap4==1.1.1
django-bulma==0.5.6
django-celery-results==1.2.1
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-debug-toolbar==1.11
django-filter==2.0.0
django-grappelli==2.14.1
django-imagekit==4.0.2
django-import-export==2.0.2
django-instagram==0.3.2
django-phonenumber-field==2.1.0
django-sb-admin==5.5.1.1
django-select2==7.1.1
django-suit==0.2.28
django-versatileimagefield==1.10
django-watermark==0.1.8
djangorestframework==3.9.2
et-xmlfile==1.0.1
facebook-sdk==3.1.0
fasttext==0.9.2
filelock==3.0.12
fonttools==3.44.0
future==0.18.2
gdown==3.11.0
geographiclib==1.50
geopy==2.0.0
google==3.0.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
httplib2==0.18.1
hyperframe==5.2.0
identify==1.5.0
idna==2.9
immutables==0.14
importlib-metadata==1.7.0
importlib-resources==3.0.0
iso-639==0.4.5
jdcal==1.4.1
Jinja2==2.11.1
joblib==0.14.1
kiwisolver==1.1.0
kombu==4.6.11
langdetect==1.0.8
lxml==4.5.2
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.0.3
mstranslate==1.1
nltk==3.5
nodeenv==1.5.0
numpy==1.18.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==2.6.4
packaging==20.4
pandas==0.24.2
phonenumbers==8.12.7
pilkit==2.0
Pillow==5.4.1
pluggy==0.13.1
pre-commit==1.21.0
prettytable==0.7.2
psycopg2-binary==2.8.4
py==1.9.0
py3dns==3.2.1
pybind11==2.5.0
pycparser==2.20
pyecharts==1.7.1
PyJWT==1.7.1
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
python-instagram==1.3.2
python-social-auth==0.3.6
python-twitter==3.5
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3
redis==3.5.3
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
simplejson==3.17.0
six==1.14.0
sniffio==1.1.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
tablib==1.1.0
textblob==0.15.3
toml==0.10.1
tox==3.20.0
tqdm==4.46.1
translate==3.5.0
Unidecode==1.1.1
urllib3==1.25.8
validate-email==1.3
vine==1.3.0
virtualenv==20.0.31
wordcloud==1.6.0
xlrd==1.2.0
XlsxWriter==1.3.3
xlwt==1.3.0
